CD: NativeZonedDateTime saving works. Next up Native Time

+++ src/main/resources/sql/oracle/JdsStoreBoolean.sql
@@ -1,7 +1,7 @@
 CREATE TABLE JdsStoreBoolean(
 	FieldId     NUMBER(19),
 	EntityGuid  NVARCHAR2(48),
-	Value       SMALLINT,
+	Value       NUMBER(3),
 	PRIMARY KEY (FieldId,EntityGuid),
 	FOREIGN KEY (EntityGuid) REFERENCES JdsStoreEntityOverview(EntityGuid) ON DELETE CASCADE
 )
