+++ src/main/resources/sql/oracle/JdsStoreEntityOverview.sql
@@ -3,6 +3,6 @@ CREATE TABLE JdsStoreEntityOverview(
     DateCreated     DATE,
     DateModified    DATE,
     Version         NUMBER(19),
-    Live            SMALLINT,
+    Live            NUMBER(3),
     PRIMARY KEY     (EntityGuid)
 )
